Split out SlackTeamCommander
